feat: clean weather widgets
